Added blocks displaying

+++ app/src/main/java/com/mars/ai1/data/repository/questions/models/AnswerOption.kt
@@ -1,7 +1,7 @@
 package com.mars.ai1.data.repository.questions.models
 
 data class AnswerOption(
-    val answerId: Long,
+    val answerId: Int,
     val name: String,
     val mark: Int
 )
